Working Android build for WebIntent Most important fixes in this commit:
(not completely verified) - when loading from WebIntent, there will be random jQuery Mobile errors, for methods such as addClass, remove, each(), etc.
This seems to be arising from the fact that the Javascript files are included in the wrong order in index.html. The  order in this index.html seemed to work so far (2-3 tries) for WebIntent template loading. I made sure that app.js loads AFTER the app is initialized, so that the document doesn't think its ready until the device is actually ready and initialized by app.initialize(); - that is my guess